3rd Place/requirements.txt
numpy==1.19.5
pandas==1.1.5
scikit-learn==0.24.2
tensorflow==2.4.1
tensorflow_addons==0.12.0
efficientnet==1.1.1
Pillow==8.2.0
ffmpeg-python==0.2.0
